% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Final EDA.R
\name{eda}
\alias{eda}
\title{Plot graphs to do Exploratory Data Analysis}
\usage{
eda(g, vect = NULL, dir = NULL)
}
\arguments{
\item{x=}{dataframe}

\item{vect=}{list of column number to be pass(if not provided, whole dataset will be taken into consideration)}

\item{dir=}{"provide valid path to save plots" (if not provided, plots will be saved in working directory)}
}
\value{
plots will be directly saved in the given directory
}
\description{
Provide Dataframe along with list of column number in the form of vector and also provide directory if you want to save plots in specific directory
(x,vect=c(1,2,3,..), dir="../EDAplots")
}
